Refactor all token sections into their own files
