+++ apps/docs/app/features/routes/ressurser/design-tokens/BreakpointTokens.tsx
@@ -1,0 +1,5 @@
+import { SharedTokenLayout } from "./SharedTokenLayout";
+
+export function BreakpointTokens() {
+  return <SharedTokenLayout title="Breakpoints"></SharedTokenLayout>;
+}
