+++ apps/docs/app/features/routes/ressurser/design-tokens/OutlineTokens.tsx
@@ -1,0 +1,5 @@
+import { SharedTokenLayout } from "./SharedTokenLayout";
+
+export function OutlineTokens() {
+  return <SharedTokenLayout title="Outlines"></SharedTokenLayout>;
+}
