+++ apps/docs/app/features/routes/ressurser/design-tokens/ZIndexTokens.tsx
@@ -1,0 +1,5 @@
+import { SharedTokenLayout } from "./SharedTokenLayout";
+
+export function ZIndexTokens() {
+  return <SharedTokenLayout title="Z-index"></SharedTokenLayout>;
+}
